Category CRUD

+++ app/src/main/java/com/saal/data/model/Category.kt
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "category")
 data class Category(
     @PrimaryKey(autoGenerate = true) @NonNull
-    var id: Long = 0L,
+    var id: Int = 0,
 
     @ColumnInfo(name = "name") @NonNull
     var name: String = ""
